Refactor models and endpoints for GoCardless integration and connection reauthorisation
- Replace `Decimal` with `float` for JSON serialization in account and transaction models.
- Update connection reauthorisation logic to support `pending` connections in addition to `expired`.
- Improve error messages and validation for connection statuses in reauthorisation endpoints.
- Add tests for `pending` connection reauthorisation and refine existing test cases.
- Remove outdated GoCardless-related PRD files after feature completion.

+++ backend/src/api/accounts/models.py
@@ -1,7 +1,6 @@
 """Pydantic models for account endpoints."""
 
 from datetime import datetime
-from decimal import Decimal
 
 from pydantic import BaseModel, Field
 
@@ -11,7 +10,8 @@ from src.postgres.common.enums import AccountStatus
 class AccountBalance(BaseModel):
     """Balance information for an account."""
 
-    amount: Decimal = Field(..., description="Balance amount")
+    # Use float for JSON serialization - Decimal would serialize as string
+    amount: float = Field(..., description="Balance amount")
     currency: str = Field(..., description="Currency code")
     type: str = Field(..., description="Balance type")
